Changing NodeInfo.getId from Integer to Long (to support all our POs having Long IDs, so I can use snowflake IDs as PKs…)

+++ src/main/java/org/code_factory/jpa/nestedset/JpaNode.java
@@ -61,7 +61,7 @@ class JpaNode<T extends NodeInfo> implements Node<T> {
         this.type = (Class<T>) node.getClass();
     }
 
-    @Override public Integer getId() {
+    @Override public Long getId() {
         return this.node.getId();
     }
